requirements.txt
streamlit
google-generativeai
gTTS
SpeechRecognition
pyaudio
pydub
pyglet